Replace latitude, longitude types in Memo.

+++ app/src/main/java/com/sap/codelab/core/data/db/MemoEntity.kt
@@ -13,7 +13,7 @@ data class MemoEntity(
     @ColumnInfo(name = "title") val title: String,
     @ColumnInfo(name = "description") val description: String,
     @ColumnInfo(name = "reminderDate") val reminderDate: Long,
-    @ColumnInfo(name = "reminderLatitude") val reminderLatitude: Long,
-    @ColumnInfo(name = "reminderLongitude") val reminderLongitude: Long,
+    @ColumnInfo(name = "reminderLatitude") val reminderLatitude: Double,
+    @ColumnInfo(name = "reminderLongitude") val reminderLongitude: Double,
     @ColumnInfo(name = "isDone") val isDone: Boolean = false
 )
